:construction_worker: Migrate the Server to FastAPI
