+++ requirements.txt
@@ -1,1 +1,4 @@
-Flask==2.1.1
+fastapi
+uvicorn
+requests
+authx
