+++ run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+set -e
+set -x
+
+uvicorn app:app --reload
